Simple refactor to the code
Breaking the filter.cs files into extension methods.
Doing just basic refactor. Once this is merged I intend to add support for nested documents filter.
