+++ FluentNest/GroupBys.cs
@@ -1,11 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Linq.Expressions;
-using Nest;
-
+﻿
 namespace FluentNest
 {
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Linq.Expressions;
+    using Nest;
+
     public static class GroupBys
     {
         /// <summary>
